requirements.txt
fastapi
python-multipart
uvicorn
httpx
pyyaml
jinja2
python-docx
pytest>=8.0.0
